Restructure MainActivity, HomeScreen and AppList
MainActivity now hosts top-level NavigableCircuitContent starting at HomepageScreen.
HomeScreen owns the VerticalSlidingDrawer and SettingsButton.
AppList is rendered as nested CircuitContent inside the drawer.
Moved settings click handler from AppListPresenter to HomepagePresenter.
Updated module dependencies accordingly.
