+++ modules/features/app-list/build.gradle.kts
@@ -33,7 +33,6 @@ android {
 
 dependencies {
     implementation(project(":modules:libs:ui"))
-    implementation(project(":modules:features:settings"))
     implementation(project(":modules:libs:intents"))
 
     implementation(platform(libs.androidx.compose.bom))
