+++ modules/features/app-list/src/test/java/com/duchastel/simon/simplelauncher/features/applist/ui/AppListPresenterTest.kt
@@ -1,9 +1,7 @@
 package com.duchastel.simon.simplelauncher.features.applist.ui
 
-import android.content.Context
 import com.duchastel.simon.simplelauncher.features.applist.data.App
 import com.duchastel.simon.simplelauncher.features.applist.data.AppRepository
-import com.duchastel.simon.simplelauncher.intents.IntentLauncher
 import com.slack.circuit.test.test
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
@@ -16,10 +14,8 @@ import org.mockito.kotlin.whenever
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class AppListPresenterTest {
-    private val context: Context = mock()
     private val appRepository: AppRepository = mock()
-    private val intentLauncher: IntentLauncher = mock()
-    private val presenter = AppListPresenter(context, appRepository, intentLauncher)
+    private val presenter = AppListPresenter(appRepository)
 
     @Test
     fun `state contains apps from repository`() = runTest {
@@ -59,14 +55,6 @@ class AppListPresenterTest {
             val state = awaitItem()
             state.apps[0].onLongClicked()
             verify(appRepository).launchAppSystemSettings(fakeApp)
-        }
-    }
-
-    fun `onSettingsClicked starts settings activity as new app`() = runTest {
-        presenter.test {
-            val state = awaitItem()
-            state.onSettingsClicked()
-            verify(intentLauncher).startActivityAsSeparateApp(any())
         }
     }
 }
