+++ modules/features/homepage/src/main/java/com/duchastel/simon/simplelauncher/features/homepage/di/HomepageModule.kt
@@ -39,7 +39,7 @@ class HomepageModule {
                     object : Ui<HomepageState> {
                         @Composable
                         override fun Content(state: HomepageState, modifier: Modifier) {
-                            Homepage(state)
+                            Homepage(state, modifier)
                         }
                     }
                 }
